Move modules/api to lib/api
